+ Some cleanup, thanks to XhmikosR
git-svn-id: svn://svn.code.sf.net/p/zenlib/code/ZenLib/trunk@420 f098ebc7-0e84-4588-8ad0-e2271d839f10

+++ Project/MSVC2005/CleanUp.bat
@@ -1,15 +1,25 @@
-@rem echo off
+@echo off
 
-@rem MS Visual Studio specific ---
-cd Example
-rmdir Win32 /Q /S
-rmdir x64/Q /S
-del *.ncb *.user
-del *.suo /AH
-cd ..
-cd Library
-rmdir Win32 /Q /S
-rmdir x64/Q /S
-del *.ncb *.user
-del *.suo /AH
-cd ..
+rem MS Visual Studio specific ---
+if exist *.opensdf del *.opensdf
+if exist *.sdf     del *.sdf
+
+pushd Example
+if exist Debug   rmdir Debug /Q /S
+if exist Release rmdir Release /Q /S
+if exist Win32   rmdir Win32 /Q /S
+if exist x64     rmdir x64 /Q /S
+if exist *.ncb   del *.ncb
+if exist *.user  del *.user
+if exist *.suo   del *.suo /AH
+popd
+
+pushd Library
+if exist Debug   rmdir Debug /Q /S
+if exist Release rmdir Release /Q /S
+if exist Win32   rmdir Win32 /Q /S
+if exist x64     rmdir x64 /Q /S
+if exist *.ncb   del *.ncb
+if exist *.user  del *.user
+if exist *.suo   del *.suo /AH
+popd
